requirements.txt
aiohttp==3.11.11
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
backoff==2.2.1
canopen==2.3.0
click==8.1.7
fastapi==0.115.6
gpiod==2.2.2
h11==0.14.0
idna==3.10
Jinja2==3.1.5
msgpack==1.1.0
packaging==24.2
psutil==6.1.0
pydantic==2.10.3
pydantic-core==2.27.1
pymodbus==3.8.1
pyserial==3.5
pysqlite3==0.5.4
python-can==4.5.0
python-dotenv==1.1.0
python-multipart==0.0.19
python-periphery==2.4.1
sniffio==1.3.1
starlette==0.41.3
typing-extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.1
wrapt==1.17.0
